+++ src/__main__.py
@@ -1,4 +1,4 @@
-"""Redirects to entrypoint of the app"""
+"""Redirects to the entrypoint of the app"""
 
 from .app import main
 from sys import exit